+++ uninstall.php
@@ -15,6 +15,7 @@ delete_option( 'lw_img_options' );
 delete_option( 'lw_img_version' );
 delete_option( 'lw_img_log' );
 delete_option( 'lw_img_bulk_job' );
+delete_option( 'lw_img_competitor_notice_dismissed' );
 
 wp_clear_scheduled_hook( 'lw_img_backup_cleanup' );
 wp_clear_scheduled_hook( 'lw_img_bulk_tick' );